+++ CHANGELOG.md
@@ -11,6 +11,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
+## [0.19.12] - 2025-12-04
+
+### Added
+
+- Added `uuidpb.FromBytes()`.
+- Added `uuidpb.UUID.AsByteArray()`.
+
 ## [0.19.11] - 2025-12-02
 
 ### Fixed
@@ -370,6 +377,7 @@ which itself includes a large numbers of breaking changes.
 [0.19.9]: https://github.com/dogmatiq/enginekit/releases/v0.19.9
 [0.19.10]: https://github.com/dogmatiq/enginekit/releases/v0.19.10
 [0.19.11]: https://github.com/dogmatiq/enginekit/releases/v0.19.11
+[0.19.12]: https://github.com/dogmatiq/enginekit/releases/v0.19.12
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD